Initial commit: title formatting tool.
This python script takes paths to mp3 files on stdin and formats their
title tags the way I like them to be. Specifically, that means:

1) Splitting the title into two parts, the song's name and then an
   optional trailing info string with any featured artists or version
   label (e.g. 'Sabrepulse Remix') that could be found. Most of the
   time, there's no info and the whole title is just the song name.

2) Formatting the name portion of the title to be in ALA+ title case,
   the plus indicating that I've added the most comon four-word
   prepositions (i.e. 'from' and 'with') to the set of words that are
   lowercased.

This script is not perfect because it performs naive word matching
instead of part of speech tagging, so it will downcase the preposition
words even when they are not being used prepositions in the title. But
it does the bulk of the work (esp. on the featured artist tags, which
can be a slog otherwise) and that's all I need it to do; I can make
minor edits the few times it does goof up.